amd/requirements.txt
panda
numpy
openai
uuid
flask==2.1.3
-e .